First attempt at cleaning styles in HQ.
Adding in blueprint css.  Deleting much of the default hq style.css
file and adding better comments and section headings to the releveant
parts.  Adding separate css files for graphs, annotations, and modal
dialogs.  Removing some old no-longer-referenced files and classes
in the css.

This is still a work in progress.