add domainsync/deidentifier app.
From readme:

This app provides the functionality to keep a domain's documents in sync
with another domain (master/slave). It also lets you apply a set of filters
/ transforms on the documents so that you can, for example, deidentify
personal information.

The original use case of this app is for creating a deidentified, "public"
domain with content that is kept in sync with an existing domain.