updating auditcare with more information to capture on view access.
adding management command to help printout all views in the project to help you figure out what views to audit.

provide better means to manage class based views, as well as updated middleware method to be useable for decorators, as decorators don't work well with the auditmiddleware, and needs to be explicitly called.