revert CommCareCase.case_count to use by_user rather than by_owner
Currently, case/by_owner has no associated reduce. It would be easy
to add, but it's not obviously the right number. I'm reverting back
to case/by_user, which will give the number of cases for which the
user has the distinction of being the last person to submit against
it.