Improve selenium testcase abstraction
Make basic config specified in class variables instead of django settings.

The only remaining dependency on django is the last fallback unittest2 import.
This is great, because there is no reason to extend from django's default
TestCase for selenium testcases, because it does a lot of time-expensive
database setup that's irrelevant for selenium.