add copy history property to domains and applications
This allows a domain to know the full chain of projects that it was copied from. In order to do this, SnapshotMixin was created which is used to help maintain the copy history for all the classes that need to store that information.

the original_doc field was removed from both applications and domains and all logic was refactored to make use of copy_history