refactor 'es_snapshot_search' to 'es_snapshot_filter'
The function is primarily filter based now and will be used specifically for filtering results. Needed to generate the facet_filters field for each facet so that the facets get filtered as well. (But it doesn't get filtered by it's own field. This is so that the client who uses the API can continue to display the unselected options for the last field selected