prevent commare properties save button from activating on application settings save
This bug took way too long to fix, for so few lines, and involves micromanaging
change events in a way that felt terrible. Also now it triggers twice sometimes,
which is fine. Next time I touch this code, I'm just going to convert it to
knockout—will take less time.