initial facilities app work
This adds a global superuser-accessible facilities app for integrating
with Facility Registry providers, integrating with the freddy API
wrapper (http://github.com/dimagi/freddy).

/facilities/registries/ provides an interface for adding, editing, and
syncing facility registries.  You can view the facilities for a
registry.

Editing a facility works and correctly sends the new data to the
facility registry server, but it has a kink that needs to get out where
extended properties get erased, and the UI needs some work.

The UI for syncing also needs some work so it's not synchronous. The
groundwork for this is already there in the form of a syncing attribute
on the model, the button just needs to be hooked in with AJAX.