More CommTrack UI-related stuff
This adds a Reminders tab that is only visible on commtrack-enabled
domains, and hides other irrelevant tabs on commtrack domains.

It also integrates all the existing settings and user pages into the
dynamically generated sidebar framework, including their usage of
"indicators" -- subpages displayed under the sidebar item only when that
subpage is currently active in order to show the user what page they are
on.

For the way I did this to work, it requires a slightly more structured
relationship between urls and views.  No two navigation items on the sidebar
should have the relationship such that one's URL is the beginning of the
other.  So I changed a couple URLs to adhere to that. (This is also true
for the relationship between tabs, subtabs, and pages, and I changed a
couple urls related to that in previous commits.)

Todo: integrate title block into this framework?