use property display for dynamic properties too (Case:56792)
display any case properties that weren't displayed in the default tables
in one below

an issue is that we need to explicitly set the th and td widths to make
it look good, but the case property names might be kind of long and lead
to ugly wrapping.

Also, we need to have really narrow tds in order to fit two groups of
two columns side-by-side on a 1280-pixel wide display, don't know what
works best here.