fix Case:46970 and Case:55261
When you delete a saved report, just reload the page to ensure that any
associated scheduled reports don't stay displayed in an outdated state
on the other tab.

When you delete a scheduled report, touch the views so
stale=update_after doesn't cause problems on the next load with the
deleted object.