Partially fix Case:54147
A while back I made the sidebar definitions declared in code in
order to make it easier to dynamically modify them depending on the
context, and so individual templates wouldn't be full of a bunch logic
to check which page is currently active.

I changed it so urls and navigation have to have a structural
isomorphism for active tabs and active menu items to be detected
correctly.  I wasn't able to test the excel import pages but it turns
out since they live at a different base URL than the data interfaces
page that you use to enter them, the navigation doesn't work with them -
when you're on one of the pages, it doesn't show Manage Data as the
selected tab, and it doesn't show anything in the sidebar.

This partially fixes that to at least show the selected tab and sidebar,
although it doesn't highlight "Import Cases from Excel" as the active
sidebar menu item, and will require extending the definition of the
navigation structure to accomodate this general case in order to achieve
that.