add left out jquery.inview.js
Missed adding this due to my over-eager global gitignore which I've now
changed.

Obviously wouldn't have passed QA if this mattered, but this is used by
the property dl pseudo-tables, which currently aren't used with more
than one column, so it's not actually necessary for the row resizing on
initial tab selection that it enables, but it's good to keep it there in
case we do enable two-column dl pseudo-tables
changed.