fix errors on form details page when form references an invalid case
the form details page used to get case ids from a view that emitted all
form ids for known cases, so this problem never came up.

Now that we look at the case ids referenced in the form, if a form has
an incorrect case id or a case was deleted due to a user being deleted
then this problem occurs.

Displaying something instead of nothing for this case is good.