Fix front end code for excel field selection
Disabling case field or custom field inputs if the other one is selected
works now. It doesn't make sense to be able to put content in both of
them.

Also made the forced formatting of input to acceptable field name happen
in real time instead of on change event firing.