add new related cases view for case details page
adds a hierarchical related cases view for the case details page.  This
is also used as the PACT Care Plan report (Case:60880).

For the care plan report, it's integrated with inline touchforms so
subcases can be added and edited at all levels from within this page.

The columns to display are configurable using the same declarative
definition format as the case details view, and the sorting of cases is
also declaratively configurable.

beginnings of configurable hierarchical related cases display