handle error with invalid unicode strings (Case:60590)
A device log was submitted containing a string that causes an error when
trying to decode to UTF-8, an invalid diacritical mark (the string was
an error message for CommCare's inability to decode this string!).

Rather than fail here we choose Python's 'replace' error-handling
method, which seems to just keep the invalid diacritic in a way that it
still gets displayed by browsers, etc.