Move default sort option code
In the original location this was persisting this default sort element.
So if you didn't sort by anything and saved the page it would show up as
a sort property the next time you load that page. In the new location it
generates the SortElement right before saving so we have it there to
write out the suite_xml but it won't be saved.