really hacky way to fix git log escaping errors
tl;dr git logs don't care about escaping, and short of doing git-rev-list or writing git alias shell scripts, this really nasty string replacement snip does the trick for just subject and message bodies.

if the parsing does fail, make a faux commit with the message so as to not blow up the deploy or the hq system admin page.