guard against jankiness
test.commcarehq.org no longer works because of this code, I presume
because the command that starts the django process isn't run from the
django root directory.

This change just guards against that.  We really should not be doing any
of this gleaning the available modules from the filesystem.

In addition, the existing report framework provides a reasonably good
interface for just making whatever sort of pages you want, and hooking
that into the URL dispatcher as an extension of the normal report
dispatching, and it also provides a central place to manage permissions
and such, so I'm a little worried that we're making a separate entry
point for other custom apps.

Even if we something separate from the report dispatcher it would be
better to have the interface be at a higher level of abstraction than
the URL router.