remove hq.legacy.js
All content of hq.legacy.js is superseded (with very few modifications)
by hqwebapp/js/main.js, which is included by some super-template of all
templates that currently include hq.legacy.js.

hq.legacy.js overwrites main.js in these cases, though, so if one of the
pages that includes it depends on the specific slightly different
functionality of the old SaveButton, that would be a problem.  I haven't
investigated or tested.

I believe this shows the differences (assuming the old SaveButton in HQ
hasn't changed much since it was copied into the formdesigner repo
(though not actually used there)):
https://github.com/mwhite/Vellum/commit/cbd6f6cfeb1ddbd5f280bad5b1c1851e3cbe75c3