Added TRAVIS_INSTALL if statement to rm #[travis-ignore] packages
The goal of this change is to have travis ignore the redis-server in requirements/apt-packages.txt which is marked with #[travis_ignore]. The grep -v command gets everything in the text file except those that contain [travis_ignore] in the line. If it's not a travis install, then the entire file is read in and installed.