Added TRAVIS_INSTALL if statement to ignore [travis_ignore] packages
The goal of this change is to have travis ignore the redis-server in requirements/apt-packages.txt which is marked with #[travis_ignore]. The grep -v command gets everything in the text file except those that contain [travis_ignore] in the line. The TRAVIS_INSTALL variable is set in travis.yml. If it's not a travis install ($TRAVIS_INSTALL != "y"), then the redis-server is read in and installed.