Show correct number of members in Group
The Group Membership page used to use group.users, which is a list of
all user ids for that group.  This includes deleted users, which was
causing the issue in http://manage.dimagi.com/default.asp?134040  This
commit uses the filtered list of users already retrieved and memoized in
the view instead.