allow disabling editing a form in vellum
This is for forms that vellum breaks, so that we can block users from
mistakenly editing them in vellum.

This feature under is a toggle (/hq/flags/edit/no_vellum/),
and is basically intended to stay there indefinitely,
so that a Dimagi power-user can disable editing certain
forms and hide that option from others
(and their future selves if they wish).

The form can still be edited by choosing Advanced > Download,
editing by hand, and reuploading with Advanced > Upload.