Fix knockout binding to Column.field
Some knockout bindings were binding to Column.field. However,
Column.field is a not a ko.observable, it is a jquery ui element.
This change adds an observable property to Column.field that
is updated whenever Column.field.val changes. This allows us to
bind to that element to stay up to date with the Column.field's
value.