Gray-out zeros in Daily Form Activity and Submissions By Form reports
(FB 147530) I'd like to highlight outliers based on median absolute
deviation. The relevant data is down columns instead of across rows,
so if the styling is done in the HTML (as opposed to calculated by the
browser) the table would need to be created twice; once for the data
and a second pass for the styling. That's a bit more code than this
little change, so I reckon we start with this, and add that as a nice
to have for later (or sooner).