(FB: 151548) - Fix MM when read-only user
When user is read-only a few errors were occurring. The first in
app-settings.html. selectionOption() would return null and then an
error would be thrown when looking .label. The next error was in
main.js when changing tabs. The unattached button had no parent element
so when checking for tagName, an error was thrown.

http://manage.dimagi.com/default.asp?151548#861201