Fix the product page pagination problems
Two things here..

1) It used to not show a loading screen even if your connection was bad
and the async call took forever. I just copied over the "searching"
concept from the users page that this is similar to. We'll want product
searching soon anyway, so no reason to come up with new names for
things.

2) Pagination used to just be wonky. Turns out that the view was sending
a string for the current page number, so javascript did '11'+3 == '113'
and weird stuff happened.