Hide less error when running Vellum in dev mode
This is a hack, but the error is a nuisance when running Vellum in dev mode (i.e., having HQ load Vellum from un-minified files in the local dev environment). I think the cause is an older version of less in HQ trying to parse newer-version less files used by Vellum. This is not a problem in production because all less files are compiled to minified CSS during the build process.