A better solution for stripping tags from Excel exports
After discussing with Cory and Daniel, this commit is a second take at
[FB 115140][1].

It still uses regex, which we would need to review if valid report values
include "<" and ">" (in that order). I've used regex because I've assumed we
don't have values like that, and regex is faster than using a full HTML parser
to strip tags.

Ultimately it would be better if cells in reports had content-type, but that
is going to be a lot more work.

This solution strips tags specifically for Excel exports, which is better than
the [previous attempt][2] which affects all exports.

  [1]: http://manage.dimagi.com/default.asp?155140
  [2]: https://github.com/dimagi/couchexport/pull/112