Fix bug with form filters
Solved by removing "--" from the column name (and replaced with "__").
"--" seems to be legal in sql columns but there is an issue in how
these names get converted to sqlalchemy bind parameters. Looks like
the Select.append_whereclause() mehtod that we are using expects
bindparameters to only include [A-Za-z0-9_].

either an issue with
sqlagg or sqlalchemy in how these column names get escaped or turned into
queries.