Avoid duplicate program lookups
The lookup will still fail hard if the program isn't found, which is
consistent with the current behavior.  It shouldn't be the case that a
product has a non-existant program_id.  I wouldn't be the least
bit surprised if that does happen, and I think I'd rather have it fail
hard and obviously when it does.