Add axis title to worker activity times report
https://developers.google.com/chart/image/
While working on this, I found that the google image chart API is deprecated.
It is only used in a few places on HQ (this, and generating QR codes). It also doesn't seem like they have plans to turn off the service yet, but it is something to look out for.