Trigger popover on hover
It seems that most help popovers are triggered by hover. Bring these in line.

Also, the group name and its icons are linked to the page to manage that group. Triggering the popover on click would require unlinking the icons. But anchor tags inside list items are styled as block-level elements, not spans. So unlinking the icons makes them appear on the next line, and look really dumb. Changing the styling affects many other places. ... Another reason to use hover instead of click.