Remove thousands separator so full number visible
One can't make the y-axis margin wider in NVD3 graphs, because of NVD3 issue 17. (Setting left margin width to 100 did not make a visible difference.) To see numbers >= 1000000 one has to remove the thousands separator. This is a poor compromise because it sacrifices readablility.

Leaving this with Review label to see whether we are happy with the compromise.

Context: [FB 169822](http://manage.dimagi.com/default.asp?169822)