Coerce group names to unicode
The xlxs library returns the same value as is stored in the spreadsheet,
so 10101 could be a string OR an int, and it's kinda opaque to the user.
Really, this parsing should be schema'd (passed to a JsonObject or
something) to validate properly, but this should do for now.

Also, `row.get` is super weird.  `row.get('name', []) == []`, but
`row.get('name', '') == None`.  Go figure.  That's why I decided not to
use that syntax.