Filter invoices by domain
http://manage.dimagi.com/default.asp?176654
Since WireInvoices and Invoices are separate models, both using Django
ORM style primary keys (integers starting at 0).  The possibility of
collisions is very high.  This means if invoice_pdf.invoice_id refers to
a WireInvoice, the first result you'd get would be an Invoice (the wrong
one).  That was causing this to 404 when the domain was checked.