Improve toggle update logic.
Note that this commit also changes which toggle caches are cleared.  The
old logic was to clear everything in the union of the old list and the
new one.  Now it only clears things that are in either the old list or
the new one, but not those in both (as these are unchanged).