Return date opened if first due visit has passed
Feeback from Matt: "I'm not sure how helpful it is to the user to see
that their visit was due a few hundred days prior and they are already a hundred days late
for their visit.  For cases that were just created, this could be the
'date_opened' for the case instead of the first eligibility window - so
that good usage would be to fill out that form on the day the case is
created (since they couldn't do any better than that).  Not sure if you
have access to date_opened, or if that is just on HQ.  (we could
manually make that case property if need be)"