Add location manager that excludes archived locs
I know I could easily just add `.filter(is_archived=False)` to the
specific instances here, but I'm doing this as a foot-in-the-door thing.
The way "archived locations" should have been implemented in the first
place is by modifying the default queryset, so archived locations are
always excluded.  My plan with this is to make a second manager and
switch stuff over piecemeal as it's identified, then eventually make it
the default.