Save form before attempting to put_attachment
There's a lot of weird interactions happening here - but I think this
missing _rev issue is related to a race condition between here and the
other handlers of the signal that gets called in the save form process.

At any rate, this was consistently reproducible locally, and this chnage
fixes it.
http://manage.dimagi.com/default.asp?182634