Add dummy cache setting so code can be loaded
I mimic what will happen on ReadTheDocs locally by doing the following:

 * Don't start my hq environment (no couch, pillowtop, redis, etc)
 * Enter my hq virtualenv
 * Move or rename `localsettings.py` so it won't be found
 * `$ cd docs/`
 * `$ make html`

Basically it needs to be able to import stuff without having a
localsettings (I just put the essentials in `dev_settings.py`) or db
connection.

In the past we've had some singleton-type objects which are
initialized on import and require a db connection to init.  These need
to either be lazily initialized (you can then memoize it to keep it a
singleton), or mocked out in `docs/conf.py`.