Quickcache UserRole.get and memoize User.role
UserRole.get is called only for non-admin users, so we devs didn't really
feel the pain of that.  On the dashboard page, I saw it get called 68
times for one page load, and for the base reports page, 144 times.

Turns out memoizing User.role was really all that was necessary.  I
figure slapping a QuickCacheDocumentMixin in here is super easy too and
doesn't hurt anything, so I did that too (it'll span multiple requests,
and I really can't imagine roles changing much).