Use raw_couch rather than doc_in_es
I was trying to find an ID recently, but got a little annoyed at the
width restrictions of the `doc_in_es` view.  Plus, it doesn't make a lot
of sense to search for these documents in elasticsearch when all we
really want is the raw JSON from couch.

I figure this is going to become more useful as we break up the couch
databases - searching in futon will become harder.