Add `mock_out_couch` decorator
It's a very thin abstraction of FakeCouchDb, but I hate working with the
mock library directly, and it's always hard to remember what to mock, so
hopefully this will encourage more people to use it.

You can use this to decorate a `SimpleTestCase` test class, and it'll
mock db calls to return nothing.  You could specify actual things to
return, but we don't use that part anywhere yet.