prefetch attachments for sql forms
This was a trick thing to figure out. For the duplicate for workflow
the form attachment is required after the form's ID has been changed
and so when you try to look them up from the DB you can't find them
since you're using the new ID.

Solution was to preftech them to a specific 'cached' attribute and
the check that list first before tyring to fetch for the db.