fix generational caching
in a previous version about django (haven’t quite pinpointed it,
but there were changes very similar to this noted in release notes
for django 1.6 and 1.7) timeout=0 meant "use the default timeout",
whereas now in 1.7 it means "expire immediately". This meant that
the code the bootstrapped the generational cache generation and
set it to 0 if it didn’t persist in the cache, and thus any new
cache would be perpetually have a generation of the original value
of 0.