Register appropriate couchapps views to new db
Notably, copy `cloudcare_apps` into couchapps, which seems like a better
home for it.  My plan is to only register this to the new db, and then
switch the calling code to this view in the flip PR.

Note that this change shouldn't result in any new views being indexed on
the main db (as I understand it, anyways).