If more than one result, try a non-fuzzy one
http://manage.dimagi.com/default.asp?215688
The filter is surprisingly complicated.

At least with the bug I was working on, the filters selected returned a
fuzzy and a non-fuzzy match.  `selected_form_data` returns None if
not precisely one match is found.  This is a strict improvement on that,
although it doesn't feel like an ideal solution.

This filtering could use a big refactor or rewrite, but I don't have
enough context and it's not clearly worth the effort at this time.