Move logging enabled to java phone settings
The ability to disable logging is meant to be a java phone-only setting, but currently appears under the general 'Advanced Settings' menu. My intention with this change is to remove it from there and it to 'Java Phone General Settings', but was just taking a guess from looking at these files, so if this doesn't do that or is missing something, let me know.