Fix abstract LogCallTestCase in wrong module
This was a very odd error. For some reason the class was getting the wrong module path when imported like

from corehq.apps.ivr.tests.util import LogCallTestCase

It was failing since this is an abstract test class with a test method that should not run until subclassed. My guess is this is caused by some kind of import weirdness, possibly a bug in Python or nose.