Alter rebuildstaging conflict resolution strategy
Whenever I see a conflict on rebuildstaging, what I usually want to know
is "does this conflict with master?" and if not, "what branches DOES" it
conflict with.  This adopts that strategy and falls back to the current
behavior if that doesn't find any conflicts.

If any objects to this change, I'd be happy to stick it behind a flag or
something.  Here's what I see now, for instance:
```
$ python scripts/rebuildstaging.py < scripts/staging.yaml --no-push rebuild
  [.] Merging vellum-staging into autostaging ok
  [.] Merging app-blobs-testing+lang-profiles into autostaging FAIL
  [.] Merging app-blobs-testing into autostaging ok
  [.] Merging lang-profiles into autostaging FAIL
  [.] Merging b3-groups into autostaging ok
  [.] Merging b3-app-multimedia into autostaging ok
  [.] Merging notifications-perms-change into autostaging ok
  [.] Merging calculatedIcons into autostaging ok
  [.] Merging check-services into autostaging ok
You must fix the following merge conflicts before rebuilding:

[.] app-blobs-testing+lang-profiles => autostaging
app-blobs-testing+lang-profiles conflicts with master

[.] lang-profiles => autostaging
lang-profiles conflicts with app-blobs-testing
```