Remove min_version
I'm pretty sure this is never relevant.  `min_version` is only ever set
if the app_id passed in corresponds to a released app, so there can't be
a situation where the latest released app has an ealier version, since
the set of released app versions *contains* the passed in app.

If I knew set theory, that could probably be proved.