[refactor] Improve reverse_agg_map
This is a refactor because when we retreive from this map, we use
a default value of "Count per choice", But really we should be
grabbing that value when "expand" is the key. It doesn't actually
matter what we return when it is "simple", because that means its
not an aggregated report.