Use jsonobject's dict coercion instead of dict()
A project somehow has an array value for user_data, and `dict()` only
does a shallow conversion, leaving the value as a jsonobject
`JsonArray`.  This change uses the `to_json()` method to more reliably
convert to basic python types, while ignoring the question of how/why
that project has complex types in the `user_data` dict in the first
place.