Add Spanish as a translatable language [ci skip]
Even when you set your preference to "es", django defaults back to English.
"Only languages listed in the LANGUAGES setting can be selected. If you want to restrict the language selection to a subset of provided languages (because your application doesn’t provide all those languages), set LANGUAGES to a list of languages."
For more, see:
https://docs.djangoproject.com/en/1.10/topics/i18n/translation/#how-django-discovers-language-preference