Fix nonresponsive select bug
This is a werid one: When trackWorkflowLink is used, there is a short period where selecting
the model type in the new export modal does nothing (i.e. doesn't cause the other fields to
appear). I think the real problem is that there is something broken in trackLinkHelper, but
I don't have the time to fix that right now.