Refer to media_files url with project id
Snapshots used to have names that matched ids, then with the
implementation of `hr_name`, the ids were truncated to form the names,
so they no longer match.  The issue at hand is that some references used
names when they should have been using ids, but only recently is that an
issue, and only for snapshots created after the `hr_name` change.
This the same issue seen here:
https://github.com/dimagi/commcare-hq/pull/13391/files