Distinguish btwn raw view and no CodeMirror
I think there IS a use-case for a totally raw json response, so this
adds that.  Additionally, it's useful to disable CodeMirror for really
large documents, where CodeMirror can be quite slow.  For that use-case,
I'd like to keep around the metadata, db statuses, and links, as
available on the normal page.