Change RelatedDoc vary_on function to account for domain
This was breaking a test.  Lookup correctly fails when the domain is bad,
then that failure is cached, so it incorrectly fails when the domain is
good.  This could conceivably happen also in production, and if a result
were cached, a malicious domain could circumvent the domain check by
hitting the cache.