Actually bulk delete locations.
Looks like this hasn't been deleting this locations properly.  I looked
at the couchdbkit code for bulk deletion, and it passes through the list
passed to it multiple times. Because `couch_locations` uses the
output of `iter_docs`, it is a generator, and after the first
pass, the generator is exhausted, so the deletion is never committed.