Allow non-superusers to import apps.
http://manage.dimagi.com/default.asp?244593
I think this was done in error.  As-is, non superusers can access the
page, but they see a blank form.  If they edit the html to remove the
"hide" class, they can submit the form sucessfully.  Server-side
validation checks for the "edit apps" permission.  This page is
documented on the public wiki, so I assume it's intended to be usable by
non superusers.