Reinstate "non numeric" language
In report builder land there are some properties that we know the
data type of, like integer questions, and some that we don't know
the type of, like case properties and hidden values.

Therefore, we know that some properties are definitely numbers,
some are unknown, and some are defintiely not numbers. "non numeric"
means we know for sure that the property is not numeric, i.e it is
a text or date question. But not "non numeric" properties could
still contain numeric data, such as case properties that always
contain numbers.