Do not save migrating_blobs_from_couch in couch
This also fixes the unfortunate circumstance where documents that were saved in couch while the flag was set on the model class were forever in a "migrating" state and would always fall back to looking for attachments in couch even after the flag had been removed from the model class. The old flag named "migrating_blobs_from_couch" will remain in couch for documents that were saved that way, but it is now ignored.