Fix migration crash on blobs != external_blobs
This state should not exist, but there is at least one app on staging that has it, and I didn't see another way to get around it. Fixing that app on staging doesn't seem worth the trouble. The owner even granted permission to delete it, although that's not trivial either.