do not add missing topics to a change feed
Because we are switching to partitioned pillows, the topics and
partitions will be determined before since. If we try to add more
topics, it will could add an intentionally ignored topic to the feed.

For example, case pillows should listen to case and case-sql. Assume
there is one partition for each and we create two pillows. The first
pillow should listen to case, second should listen to case-sql. Before
this change they would both listen to case and case-sql