I have no idea how this got broken.
This was a hell of an issue to debug.  The report wasn't showing up, no js
errors, no nothing.  Eventually I figured out that the template for the report
wasn't being rendered at all, so I found where the template is rendered (no easy
task with our view code), and saw that the base template was being rendered
directly.  With asynchronous to False, it will use the correct template.