Use is_extension() for consistency
Technically it's not possible for an owned child case that is also an extension to not already be live by the time `is_extension(owned_id)` (as added in this commit) is called. It may be slightly more efficient to check `case_id not in hosts_by_extension`: one less function call, one less hash lookup.