Split RestoreContent out of RestoreResponse
Accumulating response content in a temporary file is now managed by a separate class than serving that content in the response. Response content is no longer stored in the blob db if the content is not cached. This results in less blob db access and less temporary file churn.

Before:
- accumulate response elements in a temp file.
- put elements in another temp file wrapped in OpenRosaResponse element and delete first temp file.
- send content to blob db.
- delete second temp file.
- stream content from blob db (response).

After:
- accumulate response elements in a temp file.
- put elements in another temp file wrapped in OpenRosaResponse element and delete first temp file.
- if caching: send content to blob db.
- serve content (response).
- delete second temp file.