Use TEMP TABLE instead of CTE, also anti-join xid
After profiling it was determined that the performance of this solution is similar to the previous solution. When profiled on ICDS, the expensive part was the index scans on form_processor_commcarecaseindexsql (domain = 'icds-cas' AND case_id = cid; domain = 'icds-cas' AND referenced_id = cid; 93% of query time), which did not change with this optimization.

The second-most expensive part of the query is the join of form_processor_commcarecaseindexsql with form_processor_commcarecasesql (6% of query time), also not changed by this optimization.

It's possible that the anti-join (LEFT JOIN ... WHERE xid IS NULL) gives a small performance gain, although it was insignificant when other factors were considered.