Default to digest in auth decorators
If the default was set to "basic", which means we weren't responding with the
correct headers that are required for digest auth, including the nonce value.
This means that whenever a user tried to digest auth with any endpoint, they
never received enough information to complete their request.

Digest auth works on API endpoints as the default auth value was being set
correctly there.

From:
https://en.wikipedia.org/wiki/Digest_access_authentication#Example_with_explanation

This typical transaction consists of the following steps:

1.The client asks for a page that requires authentication but does not provide a
username and password. Typically this is because the user simply entered
the address or followed a link to the page.

2. The server responds with the 401 "Unauthorized" response code, providing the
authentication realm and a randomly generated, single-use value called a nonce.

3. At this point, the browser will present the authentication realm (typically a
description of the computer or system being accessed) to the user and prompt for
a username and password. The user may decide to cancel at this point.

4. Once a username and password have been supplied, the client re-sends the same
request but adds an authentication header that includes the response code.

5. In this example, the server accepts the authentication and the page is
returned. If the username is invalid and/or the password is incorrect, the
server might return the "401" response code and the client would prompt the user
again.