Remove a bunch of unused enikshay user_setup code
This work was started, then there were some spec changes and it was
deprioritized, and eventually I just ended up turning much of it off.  It's been
sitting around long enough that we should just remove it, and add some back in
if/when it becomes relevant again.

This should have no functional changes, note that `skip`_custom_setup currently
returns `True`, so I went through and removed everything that only happened if
that was `False`