remove second identical declaration of download_daily_saved_export
in same file. This was pretty clearly meant to be moved and not duplicated in the original
commit as well: https://github.com/dimagi/commcare-hq/commit/59a966738cb0c63446c2331379f3639114c79a83.
I believe the delete part was then blown over in a later merge commit
and didn't make it into the PR:
https://github.com/dimagi/commcare-hq/pull/14440/files