Eliminate accidental magic in tests
Missing attributes on LocationStub when standing in for SQLLocation caused the stub to be replaced with MagicMock during test validation because MockLocationStub overly broadly caught AttributeError. The comment in the AttributeError exception handler said "This will happen if the LocationStub is not new and an old_collection is not given" but that was only part of the story: it would also happen if the `old_collection` given was created by `make_collection` and contained stubs instead of real model objects.