Coerce null values to empty string
If values are set as `null`, then this probably happened in Python, and we want
to treat it as though it is an empty string, like the phone does.

This is mostly for the external_id field, as this often isn't set, but doing
queries where you compare this to the empty string don't work.